renamed files and changed directory structure
